add Sumary refactor into content and modal.
